Create a new entry point - SonarScanner.MSBuild.
