+++ Tests/SonarScanner.MSBuild.Tests/ArgumentProcessorTests.cs
@@ -24,6 +24,7 @@ using System.Linq;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarQube.Common;
+using SonarScanner.MSBuild;
 using TestUtilities;
 
 namespace SonarQube.Bootstrapper.Tests
