+++ Tests/SonarScanner.MSBuild.Tests/DefaultProcessorFactoryTests.cs
@@ -21,6 +21,7 @@
 using System;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SonarScanner.MSBuild;
 using TestUtilities;
 
 namespace SonarQube.Bootstrapper.Tests
