+++ Tests/SonarScanner.MSBuild.Tests/DefaultPropertiesFileTests.cs
