+++ Tests/SonarScanner.MSBuild.Tests/License.txt
