+++ Tests/SonarScanner.MSBuild.Tests/ProgramTests.cs
@@ -20,6 +20,7 @@
 
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using SonarScanner.MSBuild;
 using TestUtilities;
 
 namespace SonarQube.Bootstrapper.Tests
