+++ Tests/SonarScanner.MSBuild.Tests/Properties/AssemblyInfo.cs
