+++ Tests/SonarScanner.MSBuild.Tests/SonarScanner.MSBuild.Tests.ruleset
