+++ Tests/SonarScanner.MSBuild.Tests/packages.config
