+++ src/SonarScanner.MSBuild/BootstrapperClass.cs
@@ -25,7 +25,7 @@ using SonarQube.Common;
 using SonarQube.TeamBuild.Integration;
 using SonarQube.TeamBuild.Integration.Interfaces;
 
-namespace SonarQube.Bootstrapper
+namespace SonarScanner.MSBuild
 {
     public class BootstrapperClass
     {
