+++ src/SonarScanner.MSBuild/Properties/AssemblyInfo.cs
