+++ src/SonarScanner.MSBuild/Resources.resx
