+++ src/SonarScanner.MSBuild/SonarQube.Analysis.xml
